[data-import-issue] move sample data into DataImport area
