+++ packages/editor/src/components/DataImport/components/SampleData.tsx
@@ -1,21 +1,21 @@
 import React, { useContext } from 'react'
-import SampleDataContext from '../samples/SampleDataContext'
+import SampleDataContext from './samples/SampleDataContext'
 
 // Data Samples
-import validMapData from './../samples/valid-data-map.csv?raw'
-import validMapDataFootnotes from './../samples/valid-data-map-footnotes.csv?raw'
-import validChartData from './../samples/valid-data-chart.csv?raw'
-import validCountyMapData from './../samples/valid-county-data.csv?raw'
-import validGeoPoint from './../samples/valid-geo-point.csv?raw'
-import validScatterPlot from './../samples/valid-scatterplot.csv?raw'
-import validBoxPlotData from './../samples/valid-boxplot.csv?raw'
-import validAreaChart from './../samples/valid-area-chart.json?raw'
-import validWorldGeocodeData from './../samples/valid-world-geocode.json?raw'
-import validForecastData from './../samples/valid-forecast-data.csv?raw'
-import vaidWorldData from './../samples/valid-world-data.json?raw'
-import validRegionData from './../samples/valid-region-data.json?raw'
-import validSankeyData from './../samples/valid-sankey-data.json?raw'
-import pivotData from './../samples/pivotData.json?raw'
+import validMapData from './samples/valid-data-map.csv?raw'
+import validMapDataFootnotes from './samples/valid-data-map-footnotes.csv?raw'
+import validChartData from './samples/valid-data-chart.csv?raw'
+import validCountyMapData from './samples/valid-county-data.csv?raw'
+import validGeoPoint from './samples/valid-geo-point.csv?raw'
+import validScatterPlot from './samples/valid-scatterplot.csv?raw'
+import validBoxPlotData from './samples/valid-boxplot.csv?raw'
+import validAreaChart from './samples/valid-area-chart.json?raw'
+import validWorldGeocodeData from './samples/valid-world-geocode.json?raw'
+import validForecastData from './samples/valid-forecast-data.csv?raw'
+import vaidWorldData from './samples/valid-world-data.json?raw'
+import validRegionData from './samples/valid-region-data.json?raw'
+import validSankeyData from './samples/valid-sankey-data.json?raw'
+import pivotData from './samples/pivotData.json?raw'
 
 // Add additional data to samples
 const sampleData = {
@@ -99,7 +99,8 @@ const sampleData = {
 const Button = ({ text, fileName, data }) => {
   const { editingDataset, loadData } = useContext(SampleDataContext)
   const handleClick = e => loadData(new Blob([data], { type: 'text/csv' }), fileName, editingDataset)
-  const handleKeyDown = e => e.keyCode === 13 && loadData(new Blob([data], { type: 'text/csv' }), fileName, editingDataset)
+  const handleKeyDown = e =>
+    e.keyCode === 13 && loadData(new Blob([data], { type: 'text/csv' }), fileName, editingDataset)
   // prettier-ignore
   return (
     <button
@@ -113,12 +114,16 @@ const Button = ({ text, fileName, data }) => {
 
 // Map Buttons
 const MapSampleDataButtons = () => {
-  return sampleData.maps.map(sample => <Button key={sample.fileName} text={sample.text} fileName={sample.fileName} data={sample.data} />)
+  return sampleData.maps.map(sample => (
+    <Button key={sample.fileName} text={sample.text} fileName={sample.fileName} data={sample.data} />
+  ))
 }
 
 // Chart Buttons
 const ChartSampleDataButtons = () => {
-  return sampleData.charts.map(sample => <Button key={sample.fileName} text={sample.text} fileName={sample.fileName} data={sample.data} />)
+  return sampleData.charts.map(sample => (
+    <Button key={sample.fileName} text={sample.text} fileName={sample.fileName} data={sample.data} />
+  ))
 }
 
 // All Buttons
