+++ packages/editor/src/components/DataImport/components/samples/valid-area-chart.json
