+++ packages/editor/src/components/DataImport/components/samples/valid-data-map-footnotes.csv
