+++ packages/editor/src/components/DataImport/components/samples/valid-sankey-data.json
