+++ packages/editor/src/components/DataImport/components/samples/valid-scatterplot.csv
