+++ packages/editor/src/components/DataImport/components/samples/valid-world-data.json
